github: fix branch push configuration
Rename "main" to "odh". The "odh" branch is the primary branch in this
repository, so we should not configure this job to run on "main".

Remove my "ec2-github-runner-RHAIENG-2487" work-in-progress branch from
the configuration. Now that we have merged this file, we don't need this
in the configuration any more. If we want to run this workflow on other
work-in-progress branches, we can add those here at a future time.